+++ src/modify/gd/LevelSearchLayer.cpp
@@ -39,9 +39,7 @@ class $modify(LevelSearchLayer) {
 		auto enddarkhr = Mod::get()->getSettingValue<int64_t>("end-dark-hr");
 		auto enddarkmin = Mod::get()->getSettingValue<int64_t>("end-dark-min");
 	
-
 		auto screenSize = CCDirector::sharedDirector()->getWinSize();
-		float relativescale = CCDirector::sharedDirector()->getContentScaleFactor()/4;
 
 		auto DayAndNightSystem = DayAndNightSystem::create();
 		DayAndNightSystem->setID("Events"_spr);
@@ -85,12 +83,26 @@ class $modify(LevelSearchLayer) {
 
 				}
 
-				auto cc9fix = CCSprite::create("CCScale9Sprite_fix.png"_spr); // trying not to use happy textures
-        		cc9fix->setScale(1.185f * relativescale);
+				auto cc9fix = CCScale9Sprite::create("square02b_small_fix.png"_spr);
+				cc9fix->setPosition(screenSize / 2 + CCPoint {0,129.8f});
 				cc9fix->setOpacity(90);
-        		cc9fix->setPosition(screenSize / 2 + CCPoint {0,2});
-				cc9fix->setID("cc9fix"_spr);
-        		this->addChild(cc9fix,-2);
+				cc9fix->setContentSize({365,40});
+				cc9fix->setID("CC9Fix"_spr);
+				this->addChild(cc9fix,-2);
+
+				auto cc9fix2 = CCScale9Sprite::create("square02b_small_fix.png"_spr);
+				cc9fix2->setPosition(screenSize / 2 + CCPoint {0,-77.2f});
+				cc9fix2->setOpacity(90);
+				cc9fix2->setContentSize({365,50});
+				cc9fix2->setID("CC9Fix2"_spr);
+				this->addChild(cc9fix2,-2);
+				
+				auto cc9fix3 = CCScale9Sprite::create("square02b_small_fix.png"_spr);
+				cc9fix3->setPosition(screenSize / 2 + CCPoint {0,-128});
+				cc9fix3->setOpacity(90);
+				cc9fix3->setContentSize({365,35});
+				cc9fix3->setID("CC9Fix3"_spr);
+				this->addChild(cc9fix3,-2);
 			}
 		
 		}
@@ -126,12 +138,26 @@ class $modify(LevelSearchLayer) {
 					sprite_1->setOpacity(80);
 				}
 
-        		auto cc9fix = CCSprite::create("CCScale9Sprite_fix.png"_spr); // trying not to use happy textures
-        		cc9fix->setScale(1.185f * relativescale);
+        		auto cc9fix = CCScale9Sprite::create("square02b_small_fix.png"_spr);
+				cc9fix->setPosition(screenSize / 2 + CCPoint {0,129.8f});
 				cc9fix->setOpacity(90);
-        		cc9fix->setPosition(screenSize / 2 + CCPoint {0,2});
-				cc9fix->setID("cc9fix"_spr);
-        		this->addChild(cc9fix,-2);
+				cc9fix->setContentSize({365,40});
+				cc9fix->setID("CC9Fix"_spr);
+				this->addChild(cc9fix,-2);
+
+				auto cc9fix2 = CCScale9Sprite::create("square02b_small_fix.png"_spr);
+				cc9fix2->setPosition(screenSize / 2 + CCPoint {0,-77.2f});
+				cc9fix2->setOpacity(90);
+				cc9fix2->setContentSize({365,50});
+				cc9fix2->setID("CC9Fix2"_spr);
+				this->addChild(cc9fix2,-2);
+				
+				auto cc9fix3 = CCScale9Sprite::create("square02b_small_fix.png"_spr);
+				cc9fix3->setPosition(screenSize / 2 + CCPoint {0,-128});
+				cc9fix3->setOpacity(90);
+				cc9fix3->setContentSize({365,35});
+				cc9fix3->setID("CC9Fix3"_spr);
+				this->addChild(cc9fix3,-2);
 
 			}